New game
This is the base of my game so far. The game is created with javascript. The idea is that the user can control a character and a horde of zombies. The goal is to grow the horde by attacking humans. The humans actively avoid the player and horde and can also shoot at the player and the horde.